requirements.txt
ollama>=0.1.5
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
numpy>=1.24.0
python-dotenv>=1.0.0
tqdm>=4.65.0
colorama>=0.4.6
